Conflicts:
	operate/business/test/functional/BatchFreezeCouponsTest.java
	operate/business/test/functional/OperateCouponsFuncTest.java

+++ operate/business/test/functional/OperateCouponsFuncTest.java
@@ -14,7 +14,6 @@ import models.sales.Goods;
 import models.sales.Shop;
 import operate.rbac.RbacLoader;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import play.modules.paginate.JPAExtPaginator;
 import play.mvc.Http;
@@ -26,6 +25,8 @@ import java.util.Date;
 import java.util.List;
 
 /**
+ * 运营后台券功能测试.
+ *
  * User: hejun
  * Date: 12-8-22
  * Time: 上午10:29
@@ -113,7 +114,7 @@ public class OperateCouponsFuncTest extends FunctionalTest {
         Http.Response response = PUT("/coupons/" + eCoupon.id.toString() + "/freeze", "text/html", "");
         assertStatus(302, response);
         eCoupon.refresh();
-        assertEquals(1, eCoupon.isFreeze);
+        assertEquals(1, eCoupon.isFreeze.intValue());
         assertEquals(1, CouponHistory.count());
         CouponHistory historyList = CouponHistory.find("coupon=? order by createdAt desc", eCoupon).first();
         assertEquals("冻结券号", historyList.remark);
@@ -127,7 +128,7 @@ public class OperateCouponsFuncTest extends FunctionalTest {
         Http.Response response = PUT("/coupons/" + eCoupon.id.toString() + "/unfreeze", "text/html", "");
         assertStatus(302, response);
         eCoupon.refresh();
-        assertEquals(0, eCoupon.isFreeze);
+        assertEquals(0, eCoupon.isFreeze.intValue());
         assertEquals(1, CouponHistory.count());
 
         CouponHistory historyList = CouponHistory.find("coupon=? order by createdAt desc", eCoupon).first();
